Refactor wallet to use validator's record Merkle tree.
Previously, the wallet was maintaining its own copy of the record
tree which remembered the record commitments owned by the wallet,
because there was no way to instruct the validator to remember
certain records in its own tree.

This change adds a callback to validate_and_apply() which allows
the caller to control which records get remembered, and moves the
validator's existing LedgerEvent::Commit logic into this callback.
This required factoring out the wallet's owned record management into
its own data structure, to convince the borrow checker that the
closure only required access to one field of `self`. This is a big
improvement for organization and modularity anyways.