Don't persist wallet's rng state.
There is no point in doing so, since the rng is supposed to behave
as an infinite stream of entropy, and the wallet should not be able
to tell if the randomness it's getting comes directly from entropy
or from a pseudo-random, entropy-seeded stream. Therefore, it is
completely opaque to the wallet whether it is loaded with a brand
new rng or a saved rng.

Furthermore, the rng is often the only part of the dynamic state
which is changed when the wallet needs to persist itself, so
eliminating it from the persisted state saves the need for some
file system writes.