Support multiple auditor keys.
Of the 3 client key types, the auditor keys are the easiest ones
for which to make this change, hence doing them first. Following
this, we will support freezer keys and spending keys in a similar way.
See #42.

This also fixes a bug in the recent transaction history feature where
transaction history was not reverted when reverting a storage transaction.