Handle transactions (e.g. Wrap) with attached record openings.
In Spectrum, record openings are never attached to transactions;
they are always sent separately in owner memos. Thus, receiving
memos was the only way the wallet knew to add records.

In CAPE, wrap transactions have record openings attached and do
not have corresponding memos. This change adds a mechanism by which
the wallet can add record openings which were attached to a
transaction immediately when the transaction is received, without
waiting for memos.