Expose transaction building API from generic wallet.
The wallet now exposes separate interfaces for building and
submitting transactions. Allowing direct access to the transaction
builder is intended to make it easier to build applications on top
of the wallet. For example, the CAPE wallet uses features of the
transaction builder to construct burn transactions.

This also separates the concerns of building and submitting
transactions, making the design more modular. Convenience functions
are provided to build and submit transactions in one step, preserving
the old interface.

As a nice side-effect, some common parameters are wrapped up in structs
now, which fixes some clippy::too_many_arguments warnings.