Fix initial record handling in cape_ledger::text_xfr
Wallets are created with `merkle_leaf_to_forget` set to the index
of the last leaf in the Merkle tree. This is incorrect if that
leaf corresponds to an initial record belonging to the wallet, in
which case the record must not be forgotten.

Also uncommented some sync() calls needed to make the test pass
deterministically.