Streamline mock CAPE backend initialization.
Previously, there was a special case in the backend initialization
for adding initial records to wallets. This special handling led to
many bugs during development of `test_two_wallets`. Now, initial
records are handled by generating a Memos event at ledger setup time,
as in the Spectrum backend, and thus the records are added to the
wallet through the same memo handling code path as any other records.

Closes #40.