Create a framework for generalizing wallet tests.
See SpectrumXYZ/cape#107

This abstracts the extra functionality the test suite requires
from the mock backends (fault injection, etc) into an interface
that is implemented by both the Spectrum mockd and the CAPE mocks.
As a proof of concept, the `test_wallet_rejected` tests are ported
to use the new interface and thus now run with both CAPE and
Spectrum.

Further commits will port the remaining tests and remove the now-
obsolete Spectrum mocks (which have been duplicated and modified to
conform to the new interface).