Refactor api.rs so we can extract the generic parts.
The parts of api.rs that are Spectrum-specific (errors, some endpoint
return types) are moved to spectrum_api.rs. The generic parts (helpers
for dealing with tide/surf, a few AAP-general type definitions) are
still in api.rs, in preparation for extracting it into its own crate.

The main goal of this change was to redesign error handling in the
API, so instead of having one concrete Error enum, with some variants
that are Spectrum-specific, we have a trait Error, which can be
implemented differently for different APIs.