Move validator into it's own crate with library entrypoint.
This will make testing with validators easier, as you can now start
a validator using a simple Rust library call, without the need for
multiprocessing.

This is a starting point for beginning to organize our validator
codebase into a proper library, with mix-and-match support for
various functionalities including lightweight validators, full nodes,
and query services. There's still a lot of restructuring to do, but
moving the existing validator code wholesale out of `examples` and
into a proper crate is the first step.