Fix license headers
I added a bunch of new files and forgot to add licenses. Also, some
file types (e.g. HTML-style comments) were not getting automatic
license headers with the previous script.

I have changed my mind about not making license headers a precommit
hook: it is very easy to forget to add them if it's not a hook, and
there are so many files, one file with a missing license is unlikely
to be noticed. Also, the pre-commit hook is very non-intrusive, since
it's fast and automatically adds the licenses for you if they are
missing.

Thus, I have copied the license pre-commit hook from CAPE, which as
a bonus supports more file types.