Faucet: transfer sequentially
When the faucet tries to make parallel transfers, they end up getting
included in different blocks, and the finalization of the first block
invalidates the nullifier proofs in the second block. As a workaround,
this changes adds `await_transaction` calls to ensure that the faucet
only ever has one transaction in flight at a time.