Do not include nullifier proofs in transaction hash.
The nullifier proofs are auxiliary data meant to aid in validation.
They do not contribute to the fundamental identity of a transaction,
which is the effect that it has on the ledger state. Excluding them
from the hash allows validation or the EsQS to update the proofs
without confusing clients by changing the transaction hash.