Do not break up records after every faucet transfer.
The faucet is optimized to break up many records at once, by
submitting many parallel requests. Currently, we are triggering
the record breakup thread after every request, which means we only
ever break up a small number of records at a time, and we spend
about as much time breaking up records as we do serving actual
requests.

With this change, we only break up records when we drop below half
the desired number, and we replenish all the way back to the desired
number, so that we don't break up records after every request, and
when we do, we break a lot at once.