Implement metastate API using Tide Disco
The new API is the first (and, so far, only) module in the new
`full_node_esqs` module, which runs a web server alongside the
existing query service while it is under construction. Once it is
finished, we can simply remove the old `--web-server-port` parameter
and the code relating to the old query service, and fully switch
over to the new one.

Closes #357
Closes #456