Use bundled metastate api.toml if path not provided on command line.
The implementation and the api.toml are pretty tightly coupled: the
implementation is designed to work with specific route definitions
that it expects to exist in the api.toml. Generally, we should only
be running this API with the specific api.toml distributed with it.
I am keeping the configurable api.toml path as an option since it
might make debugging/prototyping easier by allowing us to tweak
api.toml without recompiling.