Fix reset-store-state logic.
The behavior involving `--reset-store-state` depends on whether
the storage directory exists or not. When we added the new
`DataSource` in `espresso-validator`, we caused the store to be
created before this flag was inspected, altering the behavior and
causing test failures. This change moves the reset-store before the
creation of the new store, and exposes a single function for doing
both in the correct order.

This also allows us to remove `--reset-store-state` from some tests,
which I had recently added to work around this.

This whole mechanism should be replaced with a simpler, better
thing that is more closely integrated with QueryData. However, that
will be easier to do once we have gotten rid of the old query service.