Integrate genesis block with espresso-validator
When including Genesis transactions in a block, in order to uphold
the invariant that `ElaboratedBlock::{block,proofs,memos}` all have
the same length, and thus avoid special-casing `Genesis` _everywhere_,
I added a `Genesis` variant of `EspressoTxnHelperProofs`, and made
`memos` optional (and bundled `Signature` with the corresponding list
of memos) so that we can use `None` as a placeholder in
`ElaboratedBlock::memos` when corresponding to a `Genesis` transaction.
This accounts for a lot of the not-directly-related changes in this
commit.