Decouple `prev_commit_time` from block height
The EsQS was using `prev_commit_time` to calculate the next block
index. This coupled block index (which should be uniquely determined
by the history) to a validation input, in addition to causing edge
cases and off-by-one errors at genesis time that confused me a lot
when dealing with the metastate bug in the previous commit.

In addition, `prev_commit_time` is supposed to refer to an opaque
notion of time that isn't necessarily the same as block height (for
example, we may soon change it to represent consensus views). This
decoupling was therefore long overdue, so this seemed as good a time
as any to do it.