Use multi-output proving keys to handle faucet requests in fewer transfers
This will decrease the number of transfers needed to satisfy a
request by up to a factor of 2. This improves the latency of a
given faucet request, and, more importantly, decreases the total
number of transfers the faucet needs to make, allowing it to sustain
a longer queue and higher demand.